restructure cart + add submit button todo
